back/pyproject.toml
[project]
name = "back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "langchain-fireworks>=0.2.1",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.31.0",
]
